requirements.txt
torch
torchvision
Pillow
opencv-python
numpy
scikit-image
scipy
scikit-learn